feat: initialize cognitive complexity
