+++ tests/CognitiveComplexity.test.ts
@@ -1,0 +1,1 @@
+import { describe, it, expect } from "@jest/globals";
