+++ tests/libs/ExtractNodes.ts
@@ -1,0 +1,10 @@
+import { describe, it, expect } from "@jest/globals";
+import { Flow } from "../../src/main/models/Flow";
+
+describe("ExtractNodes", () => {
+  it("should extract nodes from xml", () => {
+    const flow = {
+      xmldata: {},
+    } as unknown as Flow;
+  });
+});
